extracted all functions and logic from AnimatedStar to a hook
